Support index operation after resource create & modified
