+++ src/resources/resources.module.ts
@@ -3,11 +3,15 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Resource } from 'src/resources/resources.entity';
 import { ResourcesService } from 'src/resources/resources.service';
 import { ResourcesController } from 'src/resources/resources.controller';
+import { Task } from 'src/tasks/tasks.entity';
 
 @Module({
   exports: [ResourcesService],
   providers: [ResourcesService],
   controllers: [ResourcesController],
-  imports: [TypeOrmModule.forFeature([Resource])],
+  imports: [
+    TypeOrmModule.forFeature([Resource]),
+    TypeOrmModule.forFeature([Task]),
+  ],
 })
 export class ResourcesModule {}
