+++ src/tasks/tasks.controller.ts
@@ -8,9 +8,7 @@ import {
   Param,
   Post,
   Query,
-  Req,
 } from '@nestjs/common';
-import { CollectRequestDto } from 'src/tasks/dto/collect-request.dto';
 
 @Controller('api/v1/tasks')
 export class TasksController {
@@ -19,11 +17,6 @@ export class TasksController {
   @Post()
   async createTask(@Body() data: Partial<Task>) {
     return await this.tasksService.create(data);
-  }
-
-  @Post('collect')
-  async collect(@Req() req, @Body() data: CollectRequestDto) {
-    return await this.tasksService.collect(req.user, data);
   }
 
   @Get()
